refactor: Move run_cmd functions to dedicated file
This reduces the length of the main shell-interface file, improving
maintainability.
