+++ projects/butter-backup/src/butter_backup/cli.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-import contextlib
 import enum
 import json
 import os
