+++ projects/shell-interface/src/shell_interface/__init__.py
@@ -1,17 +1,19 @@
 from importlib import metadata
 
-from .shell_interface import (
+from .run_cmd import (
     PassCmdError,
     ShellInterfaceError,
+    pipe_pass_cmd_to_real_cmd,
+    run_cmd,
+)
+from .shell_interface import (
     StrPathList,
     chown,
     ensure_directory,
     get_group,
     get_user,
-    pipe_pass_cmd_to_real_cmd,
     refresh_sudo,
     rmdir_up_to,
-    run_cmd,
 )
 
 __version__ = metadata.version(__name__)
